Rename class StartScreen() --> StartMenu(), and move to folder start_screen, also fixed some minor bugs
