+++ start_screen/__init__.py
@@ -1,0 +1,1 @@
+from start_screen.start_menu import StartMenu
